Repository cleanup: consolidate docs and organize tests
Major repository organization improvements:

## Documentation Consolidation (docs/)

**Created BLS_OPTIMIZATION.md** (consolidates 6 files):
- Combines: ADAPTIVE_BLS_RESULTS, BLS_KERNEL_ANALYSIS,
  BLS_OPTIMIZATION_RESULTS, CODE_QUALITY_FIXES,
  DYNAMIC_BLOCK_SIZE_DESIGN, GPU_ARCHITECTURE_ANALYSIS
- Purpose: Single comprehensive doc for BLS performance optimization history
- Preserves: Historical context, design decisions, future opportunities
- Maintains: Technical depth while improving maintainability

**Kept relevant documentation**:
- NUFFT_LRT_README.md: User guide for Jamila Taaki's contribution
- BENCHMARKING.md: Performance benchmarking guide
- RUNPOD_DEVELOPMENT.md: Cloud GPU development workflow

**Created FILES_CLEANED.md**:
- Documents all cleanup changes
- Provides file location reference
- Lists future cleanup opportunities

**Result**: 9 markdown files → 4 (+1 cleanup doc)

## Test Organization

**Converted to proper pytest** (now in cuvarbase/tests/):

1. test_readme_examples.py (root → cuvarbase/tests/)
   - Tests README Quick Start examples work correctly
   - Verifies standard vs adaptive BLS consistency
   - 3 comprehensive test methods

2. check_nufft_lrt.py → test_nufft_lrt_import.py
   - Tests NUFFT LRT module structure and imports
   - Validates CUDA kernel existence
   - Checks documentation and examples present
   - 7 test methods

3. validation_nufft_lrt.py → test_nufft_lrt_algorithm.py
   - Tests matched filter algorithm logic (CPU-only)
   - Validates template generation, SNR computation
   - Tests perfect match, orthogonal signals, colored noise
   - 9 comprehensive test methods

**Moved to scripts/**:
- benchmark_sparse_bls.py: Benchmarks sparse BLS CPU vs GPU performance

**Deleted (redundant)**:
- test_minimal_bls.py: Nearly empty pytest stub (3 lines)
- manual_test_sparse_gpu.py: Duplicated parametrized pytest tests

**Result**: 7 Python files removed from root
- 3 converted to proper pytests in cuvarbase/tests/
- 1 moved to scripts/
- 3 deleted as redundant

## Benefits

1. **Cleaner root directory**: Only setup.py and config files remain
2. **Better test organization**: All tests are proper pytests
3. **Consolidated documentation**: Easier to maintain and find
4. **Preserved functionality**: All useful tests converted, not deleted
5. **Historical context maintained**: BLS_OPTIMIZATION.md keeps design decisions

## Testing

All tests verified working:
```bash
pytest cuvarbase/tests/test_readme_examples.py
pytest cuvarbase/tests/test_nufft_lrt_import.py
pytest cuvarbase/tests/test_nufft_lrt_algorithm.py
```
