--- check_nufft_lrt.py
@@ -1,126 +1,0 @@
-#!/usr/bin/env python
-"""
-Basic import check for NUFFT LRT module.
-This checks if the module can be imported and basic structure is accessible.
-"""
-import sys
-import os
-
-# Add current directory to path
-sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
-
-print("=" * 60)
-print("NUFFT LRT Import Check")
-print("=" * 60)
-
-# Check 1: Can we import numpy and basic dependencies?
-print("\n1. Checking basic dependencies...")
-try:
-    import numpy as np
-    print("  ✓ numpy imported successfully")
-except ImportError as e:
-    print(f"  ✗ Failed to import numpy: {e}")
-    sys.exit(1)
-
-# Check 2: Can we parse the module?
-print("\n2. Checking module syntax...")
-try:
-    import ast
-    with open('cuvarbase/nufft_lrt.py') as f:
-        ast.parse(f.read())
-    print("  ✓ Module syntax is valid")
-except Exception as e:
-    print(f"  ✗ Module syntax error: {e}")
-    sys.exit(1)
-
-# Check 3: Can we access the module structure?
-print("\n3. Checking module structure...")
-try:
-    # Try to import just to check structure (will fail if CUDA not available)
-    try:
-        from cuvarbase.nufft_lrt import NUFFTLRTAsyncProcess, NUFFTLRTMemory
-        print("  ✓ Module imported successfully (CUDA available)")
-        cuda_available = True
-    except Exception as e:
-        # This is expected if CUDA is not available
-        print(f"  ! Module import failed (CUDA not available): {e}")
-        print("  ✓ But module structure is valid")
-        cuda_available = False
-        
-except Exception as e:
-    print(f"  ✗ Unexpected error: {e}")
-    import traceback
-    traceback.print_exc()
-    sys.exit(1)
-
-# Check 4: Verify CUDA kernel exists
-print("\n4. Checking CUDA kernel...")
-try:
-    kernel_path = 'cuvarbase/kernels/nufft_lrt.cu'
-    if os.path.exists(kernel_path):
-        with open(kernel_path) as f:
-            content = f.read()
-        
-        # Count kernels
-        kernel_count = content.count('__global__')
-        print(f"  ✓ CUDA kernel file exists with {kernel_count} kernels")
-        
-        # Check for key kernels
-        required_kernels = [
-            'nufft_matched_filter',
-            'estimate_power_spectrum',
-            'compute_frequency_weights'
-        ]
-        
-        for kernel in required_kernels:
-            if kernel in content:
-                print(f"    ✓ {kernel} found")
-            else:
-                print(f"    ✗ {kernel} NOT found")
-    else:
-        print(f"  ✗ Kernel file not found: {kernel_path}")
-        sys.exit(1)
-        
-except Exception as e:
-    print(f"  ✗ Error checking kernel: {e}")
-    sys.exit(1)
-
-# Check 5: Verify tests exist
-print("\n5. Checking tests...")
-try:
-    test_path = 'cuvarbase/tests/test_nufft_lrt.py'
-    if os.path.exists(test_path):
-        with open(test_path) as f:
-            content = f.read()
-        
-        test_count = content.count('def test_')
-        print(f"  ✓ Test file exists with {test_count} test functions")
-    else:
-        print(f"  ! Test file not found: {test_path}")
-        
-except Exception as e:
-    print(f"  ! Error checking tests: {e}")
-
-# Check 6: Verify documentation exists
-print("\n6. Checking documentation...")
-try:
-    if os.path.exists('NUFFT_LRT_README.md'):
-        print("  ✓ README documentation exists")
-    else:
-        print("  ! README not found")
-        
-    if os.path.exists('examples/nufft_lrt_example.py'):
-        print("  ✓ Example code exists")
-    else:
-        print("  ! Example not found")
-        
-except Exception as e:
-    print(f"  ! Error checking documentation: {e}")
-
-print("\n" + "=" * 60)
-print("✓ All checks passed!")
-print("=" * 60)
-
-if not cuda_available:
-    print("\nNote: CUDA is not available in this environment.")
-    print("The module structure is valid and will work when CUDA is available.")
