--- docs/ADAPTIVE_BLS_RESULTS.md
@@ -1,212 +1,0 @@
-# Adaptive BLS Results
-
-## Executive Summary
-
-Dynamic block sizing provides **dramatic speedups** for small datasets, addressing the kernel-launch bottleneck identified in the baseline analysis:
-
-- **90x faster** for ndata < 64
-- **5.3x faster** for sparse ground-based surveys (ndata=100)
-- **3.4x faster** for dense ground-based surveys (ndata=500)
-- **1.4x faster** for space-based surveys (ndata=20000)
-
-**Cost savings for processing 5M lightcurves**:
-- Sparse ground-based: **$100 saved** (81% reduction)
-- Dense ground-based: **$95 saved** (71% reduction)
-- Space-based: **$114 saved** (30% reduction)
-
-## Implementation
-
-### Dynamic Block Size Selection
-
-```python
-def _choose_block_size(ndata):
-    if ndata <= 32:
-        return 32   # Single warp
-    elif ndata <= 64:
-        return 64   # Two warps
-    elif ndata <= 128:
-        return 128  # Four warps
-    else:
-        return 256  # Default (8 warps)
-```
-
-### Usage
-
-```python
-# Automatically selects optimal block size
-power = bls.eebls_gpu_fast_adaptive(t, y, dy, freqs, qmin=qmins, qmax=qmaxes)
-```
-
-## Performance Results
-
-### Synthetic Data (nfreq=1000)
-
-| ndata | Block Size | Standard (s) | Adaptive (s) | Speedup  |
-|-------|------------|--------------|--------------|----------|
-| 10    | 32         | 0.168        | 0.0018       | **93x**  |
-| 20    | 32         | 0.170        | 0.0018       | **93x**  |
-| 30    | 32         | 0.162        | 0.0018       | **90x**  |
-| 50    | 64         | 0.167        | 0.0018       | **92x**  |
-| 64    | 64         | 0.167        | 0.0018       | **93x**  |
-| 100   | 128        | 0.171        | 0.0024       | **71x**  |
-| 128   | 128        | 0.168        | 0.0025       | **67x**  |
-| 200   | 256        | 0.175        | 0.0083       | **21x**  |
-| 500   | 256        | 0.166        | 0.0366       | **4.5x** |
-| 1000  | 256        | 0.172        | 0.0708       | **2.4x** |
-| 5000  | 256        | 0.180        | 0.1646       | **1.1x** |
-| 10000 | 256        | 0.176        | 0.1747       | **1.0x** |
-
-### Realistic Keplerian BLS (10-year baseline)
-
-#### Sparse Ground-Based (ndata=100, nfreq=480k)
-- Standard: 0.260s per lightcurve
-- Adaptive: 0.049s per lightcurve
-- **Speedup: 5.33x**
-- Cost for 5M LCs: $123 → $23 (**$100 saved, 81% reduction**)
-
-#### Dense Ground-Based (ndata=500, nfreq=734k)
-- Standard: 0.283s per lightcurve
-- Adaptive: 0.082s per lightcurve
-- **Speedup: 3.44x**
-- Cost for 5M LCs: $134 → $39 (**$95 saved, 71% reduction**)
-
-#### Space-Based (ndata=20k, nfreq=891k)
-- Standard: 0.797s per lightcurve
-- Adaptive: 0.554s per lightcurve
-- **Speedup: 1.44x**
-- Cost for 5M LCs: $376 → $262 (**$114 saved, 30% reduction**)
-
-## Analysis
-
-### Why Such Dramatic Speedups?
-
-The baseline analysis identified ~0.17s constant kernel launch overhead. For small ndata:
-
-**Before (block_size=256)**:
-- Thread utilization: 10/256 = 3.9% for ndata=10
-- Most threads idle
-- 0.17s overhead + minimal compute
-
-**After (block_size=32)**:
-- Thread utilization: 10/32 = 31% for ndata=10
-- 8x fewer idle threads
-- Kernel launches much faster
-- 0.0018s total time!
-
-### Speedup vs ndata
-
-The speedup curve shows clear regions:
-
-1. **ndata < 64**: 90x+ speedup
-   - Block size 32-64
-   - Kernel launch overhead eliminated
-   - Throughput increased from 0.06 to 5-35 M eval/s
-
-2. **64 < ndata < 200**: 20-70x speedup
-   - Block size 128
-   - Still significant launch overhead reduction
-
-3. **200 < ndata < 1000**: 2-20x speedup
-   - Block size 256 (same as baseline)
-   - But with optimized kernel (bank conflicts fixed)
-   - Reduced overhead from better utilization
-
-4. **ndata > 1000**: ~1x speedup
-   - Block size 256
-   - Already compute-bound, not launch-bound
-   - As expected from initial analysis
-
-### Real-World Impact
-
-For typical survey use cases, the adaptive approach provides:
-
-**Sparse ground-based surveys** (HAT, MEarth, NGTS):
-- ~100-500 observations per lightcurve
-- 5-90x faster processing
-- 71-81% cost reduction
-- **Enables affordable all-sky BLS searches**
-
-**Dense space-based surveys** (TESS, Kepler):
-- ~20k observations per lightcurve
-- 1.4x faster processing
-- 30% cost reduction
-- **Still significant savings at scale**
-
-## Correctness Verification
-
-All block sizes produce identical results within floating-point precision:
-- Max difference: < 3e-8
-- Typical difference: 0 (exact match)
-- Verified across all test configurations
-
-## Comparison to Previous Optimizations
-
-| Optimization                  | ndata=10 | ndata=100 | ndata=1000 | ndata=10k |
-|-------------------------------|----------|-----------|------------|-----------|
-| Baseline (block_size=256)     | 1.00x    | 1.00x     | 1.00x      | 1.00x     |
-| Bank conflict fix + shuffles  | 1.05x    | 0.97x     | 1.06x      | 0.98x     |
-| **Adaptive block sizing**     | **93x**  | **71x**   | **2.4x**   | **1.0x**  |
-
-The adaptive approach provides **1-2 orders of magnitude** better speedup than micro-optimizations by addressing the actual bottleneck.
-
-## Recommendations
-
-### For Users
-
-**Use `eebls_gpu_fast_adaptive()` by default**:
-```python
-# Replaces eebls_gpu_fast()
-power = bls.eebls_gpu_fast_adaptive(t, y, dy, freqs, qmin=qmins, qmax=qmaxes)
-```
-
-**When to use standard version**:
-- Never! Adaptive is strictly better or equal
-- Falls back to block_size=256 for large ndata anyway
-
-### For Batch Processing
-
-The adaptive approach is **especially beneficial** for batch processing:
-
-```python
-# Process 1000 lightcurves
-for t, y, dy in lightcurves:
-    power = bls.eebls_gpu_fast_adaptive(t, y, dy, freqs, qmin=qmins, qmax=qmaxes)
-    # 5-90x faster than standard!
-```
-
-Kernel caching ensures no compilation overhead for repeated calls.
-
-### Future Work
-
-Potential further improvements:
-
-1. **Frequency batching** for very small ndata
-   - Process multiple frequencies in single kernel launch
-   - Could provide additional 2-5x for ndata < 20
-
-2. **Stream batching** for multiple lightcurves
-   - Launch multiple lightcurves in parallel streams
-   - Overlap compute with memory transfer
-   - Could provide 1.5-2x throughput improvement
-
-3. **Persistent kernels**
-   - Avoid kernel launch entirely
-   - Keep GPU continuously busy
-   - Most complex but highest potential (10x+)
-
-## Conclusion
-
-Dynamic block sizing successfully addresses the kernel-launch bottleneck:
-
-- ✅ **90x speedup** for small datasets (ndata < 64)
-- ✅ **5x speedup** for typical ground-based surveys
-- ✅ **Zero regression** for large datasets
-- ✅ **Automatic** - no user intervention needed
-- ✅ **Production-ready** - verified correctness
-
-This represents the **single most impactful optimization** for BLS performance, providing:
-- **$100-200 cost savings** per 5M lightcurves
-- **10-100x faster** batch processing for sparse surveys
-- **Enables previously infeasible** all-sky BLS searches
-
-The implementation is clean, maintainable, and backward-compatible, making it suitable for immediate adoption in production pipelines.
