--- docs/BLS_OPTIMIZATION_RESULTS.md
@@ -1,127 +1,0 @@
-# BLS Kernel Optimization Results
-
-## Summary
-
-Implemented and tested an optimized version of the BLS CUDA kernel with the following improvements:
-- Fixed bank conflicts (separate yw/w arrays)
-- Fast math intrinsics (`__float2int_rd`, `mod1_fast`)
-- Warp shuffle reduction (eliminates 4 `__syncthreads` calls)
-
-## Performance Results
-
-Benchmarked on RTX 4000 Ada Generation with nfreq=1000, 5 trials per configuration:
-
-| ndata  | Standard (s) | Optimized (s) | Speedup | Max Diff     |
-|--------|--------------|---------------|---------|--------------|
-| 10     | 0.1704       | 0.1793        | 0.95x   | 0.00e+00     |
-| 100    | 0.1710       | 0.1759        | 0.97x   | 2.98e-08     |
-| 1000   | 0.1728       | 0.1625        | 1.06x   | 1.12e-08     |
-| 10000  | 0.1723       | 0.1758        | 0.98x   | 5.59e-09     |
-
-**Key Finding**: Only modest improvements (6% speedup at best for ndata=1000), with no improvement or slight slowdowns in other cases.
-
-## Correctness Verification
-
-Optimized kernel produces results within floating-point precision of standard kernel:
-- Max absolute difference: 7.45e-09
-- Max relative difference: 3.33e-07
-- Well within acceptable tolerance (< 1e-4)
-
-## Analysis
-
-### Why Limited Speedup?
-
-The baseline analysis identified that the kernel is **kernel-launch bound** rather than compute-bound:
-- Runtime is nearly constant (~0.17s) regardless of ndata
-- For ndata=10: only 10/256 = 3.9% thread utilization
-- Kernel launch overhead dominates for small ndata
-
-Our optimizations addressed compute-side bottlenecks (bank conflicts, reduction algorithm), but these weren't the limiting factor.
-
-### What Would Actually Help?
-
-Based on the analysis, significant speedups would require:
-
-1. **Dynamic block sizing** (5x potential for small ndata)
-   - Use smaller blocks for small ndata
-   - Batch multiple frequencies per block
-   - This would address the 3.9% utilization issue
-
-2. **Reduced kernel launch overhead**
-   - Stream batching
-   - Persistent kernels
-   - These address the constant ~0.15s baseline
-
-3. **Memory access improvements** (30% potential)
-   - Texture memory for read-only data
-   - Better coalescing patterns
-
-### What We Did Achieve
-
-While speedups were modest, the optimizations are still valuable:
-
-1. **No performance regression** - within noise for most cases
-2. **Numerically identical results** - differences < 1e-7
-3. **Better code quality**:
-   - Eliminated bank conflicts (cleaner memory access)
-   - More efficient warp-level primitives
-   - Explicit use of fast math (compiler flag was already set)
-4. **Established benchmark infrastructure** for future work
-
-## Implementation Details
-
-### Files Modified
-- `cuvarbase/kernels/bls_optimized.cu` - New optimized kernel
-- `cuvarbase/bls.py` - Added `eebls_gpu_fast_optimized()` and `use_optimized` parameter
-- `scripts/compare_bls_optimized.py` - Comparison benchmark
-- `scripts/test_optimized_correctness.py` - Correctness verification
-
-### Key Bug Fixed During Development
-
-Initial version had a critical bug in the warp shuffle reduction:
-```cuda
-// WRONG: Stops before handling k=32 case
-for(unsigned int k = (blockDim.x / 2); k > 32; k /= 2)
-
-// CORRECT: Includes k=32 iteration
-for(unsigned int k = (blockDim.x / 2); k >= 32; k /= 2)
-```
-
-This caused the optimized kernel to produce incorrect results (up to 65% relative error) until fixed.
-
-## Recommendations
-
-### For Users
-- Use standard `eebls_gpu_fast()` - the optimized version offers minimal benefit
-- Optimized version available via `eebls_gpu_fast_optimized()` for testing
-
-### For Future Development
-
-Priority optimizations for meaningful speedup:
-
-1. **HIGH PRIORITY**: Implement dynamic block sizing
-   - Detect ndata and adjust block size accordingly
-   - For ndata < 100: use 32 or 64 thread blocks
-   - For ndata > 1000: keep 256 thread blocks
-   - Batch frequencies for small ndata cases
-
-2. **MEDIUM PRIORITY**: Implement texture memory for t, yw, w arrays
-   - All blocks read same data
-   - Texture cache would benefit repeated access
-   - Expected 10-20% improvement
-
-3. **LOW PRIORITY**: Atomic operation reduction
-   - Private histograms per warp
-   - Warp-level reduction before atomics
-   - Most beneficial for large ndata (> 10k)
-
-## Conclusion
-
-This optimization effort successfully:
-- ✓ Implemented production-quality optimized kernel
-- ✓ Verified numerical correctness
-- ✓ Identified kernel-launch bottleneck as true limiting factor
-- ✓ Established benchmark infrastructure
-- ✓ Documented clear path for future improvements
-
-While speedups were modest (< 10%), the work provides a solid foundation for more impactful optimizations targeting the actual bottleneck (kernel launch overhead and thread utilization).
