+++ docs/FILES_CLEANED.md
@@ -1,0 +1,180 @@
+# Repository Cleanup Summary
+
+**Date**: October 2025
+**Branch**: `repository-cleanup`
+
+This document summarizes the repository cleanup performed to consolidate documentation and organize test files.
+
+---
+
+## Markdown Documentation (docs/)
+
+### Files Kept
+
+1. **BLS_OPTIMIZATION.md** (NEW - consolidates 6 old files)
+   - **Purpose**: Chronicles all BLS GPU performance optimizations
+   - **Content**: Adaptive block sizing (90x speedup), micro-optimizations, thread-safety
+   - **Historical**: Documents optimization decisions and future opportunities
+   - **For**: Developers interested in performance improvements and future optimization work
+
+2. **NUFFT_LRT_README.md**
+   - **Purpose**: Documentation for NUFFT-based Likelihood Ratio Test
+   - **Content**: Algorithm explanation, usage examples, API reference, citations
+   - **Credits**: Jamila Taaki's contribution
+   - **For**: Users wanting to use NUFFT-LRT for transit detection with correlated noise
+
+3. **BENCHMARKING.md**
+   - **Purpose**: Guide for running performance benchmarks
+   - **Content**: Instructions, example results, interpretation
+   - **For**: Developers benchmarking performance or comparing algorithms
+
+4. **RUNPOD_DEVELOPMENT.md**
+   - **Purpose**: Workflow for developing locally with cloud GPU testing
+   - **Content**: RunPod setup, sync scripts, remote testing
+   - **For**: Developers without local GPUs who need to test on cloud instances
+
+### Files Removed (Consolidated into BLS_OPTIMIZATION.md)
+
+- ❌ **ADAPTIVE_BLS_RESULTS.md** - Detailed adaptive BLS benchmark results
+- ❌ **BLS_KERNEL_ANALYSIS.md** - Baseline profiling and bottleneck analysis
+- ❌ **BLS_OPTIMIZATION_RESULTS.md** - Micro-optimization benchmark results
+- ❌ **CODE_QUALITY_FIXES.md** - Thread-safety and LRU cache implementation
+- ❌ **DYNAMIC_BLOCK_SIZE_DESIGN.md** - Design document for adaptive block sizing
+- ❌ **GPU_ARCHITECTURE_ANALYSIS.md** - GPU scaling and batching analysis
+
+**Rationale**: Too many docs for a single feature. Consolidated into one comprehensive document that preserves historical context while being more maintainable.
+
+---
+
+## Top-Level Python Scripts
+
+### Files Kept
+
+1. **setup.py**
+   - **Purpose**: Package installation script (required)
+   - **Status**: Must keep for `pip install`
+
+### Files Converted to pytest
+
+2. **test_readme_examples.py** → `cuvarbase/tests/test_readme_examples.py`
+   - **Purpose**: Tests that README code examples work correctly
+   - **New location**: Proper pytest in test suite
+   - **Tests**: Quick Start example, standard vs adaptive BLS consistency
+
+3. **check_nufft_lrt.py** → `cuvarbase/tests/test_nufft_lrt_import.py`
+   - **Purpose**: Validates NUFFT LRT module structure and imports
+   - **New location**: Proper pytest for module structure validation
+   - **Tests**: Syntax validation, CUDA kernel existence, documentation presence
+
+4. **validation_nufft_lrt.py** → `cuvarbase/tests/test_nufft_lrt_algorithm.py`
+   - **Purpose**: Tests matched filter algorithm logic (CPU-only)
+   - **New location**: Proper pytest for algorithm validation
+   - **Tests**: Template generation, perfect match, orthogonal signals, scale invariance, colored noise
+
+### Files Moved to scripts/
+
+5. **benchmark_sparse_bls.py** → `scripts/benchmark_sparse_bls.py`
+   - **Purpose**: Benchmarks sparse BLS CPU vs GPU performance
+   - **New location**: Consolidated with other benchmark scripts in `scripts/`
+
+### Files Deleted (Redundant)
+
+- ❌ **test_minimal_bls.py** - Nearly empty pytest stub (3 lines)
+- ❌ **manual_test_sparse_gpu.py** - Redundant with `test_bls.py::test_sparse_bls_gpu`
+
+**Rationale**:
+- `test_minimal_bls.py` had no real tests
+- `manual_test_sparse_gpu.py` duplicated existing parametrized pytest tests
+
+---
+
+## Summary of Changes
+
+### Documentation
+- **Before**: 9 markdown files in `docs/`
+- **After**: 4 markdown files in `docs/`
+- **Net**: -5 files (consolidated 6 into 1, kept 3)
+
+### Top-Level Scripts
+- **Before**: 7 Python files in root (excluding `setup.py`)
+- **After**: 0 Python files in root (excluding `setup.py`)
+- **Net**: -7 files from root
+  - 3 converted to proper pytests in `cuvarbase/tests/`
+  - 1 moved to `scripts/`
+  - 3 deleted (redundant)
+
+### Benefits
+1. **Cleaner root directory**: Only `setup.py` and configuration files remain
+2. **Better test organization**: All tests are proper pytests in `cuvarbase/tests/`
+3. **Consolidated documentation**: Easier to maintain, find, and update
+4. **Preserved context**: BLS_OPTIMIZATION.md keeps historical optimization decisions
+5. **No functionality lost**: All useful tests converted to pytest, not deleted
+
+---
+
+## File Locations Reference
+
+### Documentation (docs/)
+```
+docs/
+├── BLS_OPTIMIZATION.md          # BLS performance optimization history
+├── NUFFT_LRT_README.md           # NUFFT-LRT user guide
+├── BENCHMARKING.md               # Benchmarking guide
+└── RUNPOD_DEVELOPMENT.md         # Cloud GPU development workflow
+```
+
+### Tests (cuvarbase/tests/)
+```
+cuvarbase/tests/
+├── test_readme_examples.py       # Tests README code examples
+├── test_nufft_lrt_import.py      # Tests NUFFT LRT module structure
+└── test_nufft_lrt_algorithm.py   # Tests NUFFT LRT algorithm logic (CPU)
+```
+
+### Scripts (scripts/)
+```
+scripts/
+├── benchmark_sparse_bls.py       # Benchmark sparse BLS performance
+├── benchmark_adaptive_bls.py      # Benchmark adaptive BLS
+├── benchmark_algorithms.py        # General algorithm benchmarks
+└── ... (other existing scripts)
+```
+
+---
+
+## Testing After Cleanup
+
+To verify all tests still work:
+
+```bash
+# Run all tests
+pytest cuvarbase/tests/
+
+# Run specific test files
+pytest cuvarbase/tests/test_readme_examples.py
+pytest cuvarbase/tests/test_nufft_lrt_import.py
+pytest cuvarbase/tests/test_nufft_lrt_algorithm.py
+```
+
+To run benchmarks:
+
+```bash
+# Sparse BLS benchmark
+python scripts/benchmark_sparse_bls.py
+
+# Adaptive BLS benchmark
+python scripts/benchmark_adaptive_bls.py
+```
+
+---
+
+## Future Cleanup Opportunities
+
+Items not addressed in this cleanup (can be done later if needed):
+
+1. **copilot-generated/** directory in docs/ - Contains old Copilot-generated documentation
+2. **analysis/** directory in root - Contains TESS cost analysis scripts
+3. **examples/benchmark_results/** - Old benchmark results (could archive or remove)
+4. **.json files in root** - Benchmark result files (`standard_bls_benchmark.json`, `tess_cost_analysis.json`)
+
+These were not cleaned up in this pass to stay focused on the immediate goals (consolidate docs, organize tests).
