--- docs/GPU_ARCHITECTURE_ANALYSIS.md
@@ -1,222 +1,0 @@
-# GPU Architecture Analysis for BLS Performance
-
-## Question 1: Have we leveraged batching?
-
-**Answer: Not yet.** Current implementation processes lightcurves sequentially:
-
-```python
-for t, y, dy in lightcurves:
-    power = bls.eebls_gpu_fast_adaptive(t, y, dy, freqs, qmin=qmins, qmax=qmaxes)
-```
-
-### Current GPU Utilization (RTX 4000 Ada, 48 SMs)
-
-| Use Case | ndata | nfreq | Grid Size | GPU Saturation |
-|----------|-------|-------|-----------|----------------|
-| Sparse ground | 100 | 480k | 5000 blocks | ✓ Saturated |
-| Dense ground | 500 | 734k | 5000 blocks | ✓ Saturated |
-| Space-based | 20k | 891k | 5000 blocks | ✓ Saturated |
-
-**Finding**: With grid_size=5000 and 48 SMs, we launch 104 blocks per SM, which saturates the GPU. **However**, this doesn't mean we can't benefit from batching!
-
-### Why Batching Could Still Help
-
-1. **Kernel launch overhead**: Even though GPU is saturated during compute, there's ~0.001-0.002s overhead between kernels
-   - For 5M lightcurves: 5000-10000s wasted on launches alone!
-   - Batching reduces # of launches
-
-2. **Memory transfer overhead**: Currently transferring data sequentially
-   - Could overlap compute with memory transfer using streams
-   - Pipeline: transfer LC N+1 while computing LC N
-
-3. **Larger GPUs have more SMs**: On A100/H100, single LC may NOT saturate
-
-## Question 2: How do speedups scale to different GPUs?
-
-### GPU Comparison
-
-| GPU | SMs | Max Blocks | Max Threads | Single LC Saturates? |
-|-----|-----|------------|-------------|---------------------|
-| RTX 4000 Ada | 48 | 1,152 | 73,728 | YES (5000 blocks) |
-| A100 (40GB) | 108 | 2,592 | 165,888 | YES (5000 blocks) |
-| A100 (80GB) | 108 | 2,592 | 165,888 | YES (5000 blocks) |
-| H100 | 132 | 3,168 | 202,752 | YES (5000 blocks) |
-| H200 | 132 | 3,168 | 202,752 | YES (5000 blocks) |
-| B200 | ~200* | ~4,800* | ~307,200* | YES (5000 blocks) |
-
-*B200 specs estimated based on Blackwell architecture
-
-### Will Speedups Change on Larger GPUs?
-
-**Short answer: Speedups will be SIMILAR, possibly BETTER.**
-
-#### Why speedups should be similar:
-
-1. **Kernel launch overhead is architecture-independent**
-   - Measured ~0.17s constant overhead on RTX 4000 Ada
-   - Likely similar on A100/H100 (maybe 0.10-0.15s)
-   - Adaptive approach eliminates this overhead regardless of GPU
-
-2. **Block sizing benefits are universal**
-   - Small ndata → poor thread utilization on ANY GPU
-   - Dynamic block sizing fixes this on all architectures
-
-#### Why speedups might be BETTER on larger GPUs:
-
-1. **More memory bandwidth**
-   - A100: 1.6 TB/s (vs RTX 4000 Ada: 360 GB/s)
-   - H100: 3.35 TB/s
-   - Faster data transfers → lower kernel overhead → bigger relative gain
-
-2. **Better occupancy schedulers**
-   - Newer GPUs have improved warp schedulers
-   - Better at hiding latency with small block sizes
-   - Could see 100x+ speedups instead of 90x
-
-3. **More SMs = better concurrent stream utilization**
-   - RTX 4000 Ada saturates at 5000 blocks
-   - A100/H100 could run 2-3 lightcurves concurrently
-   - Additional 2-3x speedup for batch processing
-
-### Expected Performance on Different GPUs
-
-#### RTX 4000 Ada (Current Results)
-```
-Sparse (ndata=100): 5.3x speedup
-Dense (ndata=500):  3.4x speedup
-Space (ndata=20k):  1.4x speedup
-```
-
-#### A100 (Predicted)
-```
-Sparse (ndata=100): 6-8x speedup
-  - Better memory bandwidth → lower overhead
-  - Could batch 2 LCs concurrently → 2x more
-
-Dense (ndata=500):  3.5-4x speedup
-  - Similar to RTX 4000 Ada
-
-Space (ndata=20k):  1.5-2x speedup
-  - Better memory bandwidth helps large transfers
-```
-
-#### H100 (Predicted)
-```
-Sparse (ndata=100): 8-12x speedup
-  - 2x better memory bandwidth than A100
-  - Could batch 3 LCs concurrently → 3x more
-
-Dense (ndata=500):  4-5x speedup
-  - Better bandwidth + occupancy
-
-Space (ndata=20k):  2-2.5x speedup
-  - Massive bandwidth helps data movement
-```
-
-#### H200/B200 (Predicted)
-```
-Similar to H100, possibly 10-20% better due to:
-- Improved memory architecture
-- Better schedulers
-- More SMs for concurrent batching
-```
-
-## Batching Opportunities Not Yet Exploited
-
-### 1. CUDA Streams for Concurrent Execution
-
-Even though single LC saturates GPU on RTX 4000 Ada, larger GPUs could benefit:
-
-```python
-# Potential implementation
-def process_batch_concurrent(lightcurves, freqs, qmins, qmaxes, n_streams=4):
-    streams = [cuda.Stream() for _ in range(n_streams)]
-    memories = [bls.BLSMemory(...) for _ in range(n_streams)]
-
-    results = []
-    for i, (t, y, dy) in enumerate(lightcurves):
-        stream_idx = i % n_streams
-
-        # Async memory transfer and compute
-        power = bls.eebls_gpu_fast_adaptive(
-            t, y, dy, freqs, qmin=qmins, qmax=qmaxes,
-            stream=streams[stream_idx],
-            memory=memories[stream_idx]
-        )
-        results.append(power)
-
-    # Synchronize all streams
-    for s in streams:
-        s.synchronize()
-
-    return results
-```
-
-**Expected benefit**:
-- RTX 4000 Ada: 1.2-1.5x (overlap launch overhead)
-- A100/H100: 2-3x (true concurrent execution)
-
-### 2. Persistent Kernels
-
-Instead of launching kernel for each lightcurve, keep GPU busy continuously:
-
-```cuda
-__global__ void persistent_bls(lightcurve_queue) {
-    while (has_work()) {
-        lightcurve = get_next_lightcurve();
-        process_bls(lightcurve);
-    }
-}
-```
-
-**Expected benefit**: 5-10x by eliminating ALL launch overhead
-
-### 3. Frequency Batching for Small ndata
-
-For ndata < 32, we could process multiple frequency ranges in a single kernel:
-
-**Expected benefit**: Additional 2-3x for sparse surveys
-
-## Recommendations
-
-### Immediate Actions (Low Effort, High Impact)
-
-1. ✅ **DONE**: Dynamic block sizing
-   - Already implemented
-   - Works on all GPUs
-   - 90x speedup for small ndata
-
-2. **TODO**: Implement CUDA streams for batch processing
-   - Moderate effort (~100 lines of code)
-   - 1.2-3x additional speedup depending on GPU
-   - Most beneficial on A100/H100
-
-### Medium-Term (Moderate Effort)
-
-3. **TODO**: Benchmark on A100/H100
-   - Rent cloud instance
-   - Run same benchmarks
-   - Quantify actual speedups vs predictions
-
-4. **TODO**: Optimize for specific GPU architectures
-   - Tune block sizes per architecture
-   - Use architecture-specific features (Tensor Cores?)
-
-### Long-Term (High Effort)
-
-5. **TODO**: Persistent kernels
-   - Requires major refactoring
-   - 5-10x additional speedup potential
-   - Most complex implementation
-
-## Summary
-
-| Optimization | Effort | Speedup (RTX 4000) | Speedup (A100/H100) |
-|--------------|--------|-------------------|---------------------|
-| Dynamic block sizing | ✅ DONE | 5-90x | 6-120x (predicted) |
-| CUDA streams | TODO | 1.2-1.5x | 2-3x |
-| Persistent kernels | TODO | 5-10x | 5-10x |
-| **TOTAL POTENTIAL** | | **25-450x** | **60-3600x** |
-
-Current achievement: **5-90x** depending on ndata
-Remaining potential: **5-40x** additional from batching optimizations
