--- test_minimal_bls.py
@@ -1,6 +1,0 @@
-import pytest
-from cuvarbase.bls import sparse_bls_gpu, compile_bls, eebls_gpu
-
-def test_minimal():
-    """Minimal test"""
-    pass
